beginning of refactoring. New features, and scaffolding for even more.
** GENERATE.PHP SHOULD NOT BE USED AT THIS POINT! ***

The get_metadata() function was getting too unwieldy, it has been split
up into:
  * get_metadata(), which grabs stuff out of the file and from the
    filesystem, and calls
  * process_metadata(), which expands the $metadata array based on
    things discovered during get_metadata(),
  * blog_order(), whose sole purpose is to order $metadata into reverse
    chronological order.

Adding arbitrary key/value pairs to the METADATA section of the file
will add them to the $metadata array. It is hoped this will allow a new
templating system to use these as necessary, without the need to change
the generating scripts.

Preliminary support for the webkit2png program, which uses command-line
webkit to make a PNG of a webpage. Posts of 'link' type will cause a
'webshot' to be generated in the /images/webshots/ directory, and an img
tag to be added to the post_content metadata. If the shot already
exists, it will not be recreated. The filename for these images is based
on the 'posted' filename, so no collisions should occur.  A
make_webshot() function exists for creating/managing these webshots, and
a $webkit2png variable has been added to config.php, holding the path to
the webkit2png script and sizing options.

Support for 'pre_content' and 'post_content' settings in the metadata
array. These are not yet used in the templates.

Posts of 'quote' type wraps the post in a div (or, rather, will when
support for pre_content and post_content make it into the templates).

Posts of 'image' type adds the image to the pre_content metadata. This
allows the post to act as the caption for the image. In the future,
these images (whose url is in the 'url' metadata) will be cached
locally.

Posts of 'link' type will now link the title of the post directly to the
link target.

Some more E_ warnings have been fixed.

The test.php file has been added for diagnostics during the refactoring.