We'll call this one-inch-frame 1.1.
'source' now an officially supported piece of metadata. If found, it
will be linked from the webshot for type='link', and the image for
type='image'. In addition, it will be added to a "sources" ul in
post_content as "(source)".

The 'url' metadata is likewise added to the 'sources' ul, as "(view
original)".

The twitter block has been added to the sidebar of index.html.

It is recommended that templates that have a <head> section (that is, a
top-level template) should include the line:

  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />

to best allow special characters like smart quotes and emdashes to be
output to the web.

This is ready to produce web sites again.