Separated webshots out into a separate script that is called by make_webshot().
Now, instead of needing the full path and options to webkit2png,
config.php contains a variable pointing to a shell script which accepts
the necessary parameters and creates the PNG for the webshot. In this
way, users have full control over the shot created (except for the file
name and path). An example script, which I use on eafarris.com, is
included as webshot.sh. It uses webkit2png to grab the png, then uses
ImageMagick to manipulate it. It could do other things like recolor,
blur, etc. And, if there's a better tool out there besides webkit2png, I
can switch to it without changing the make_webshot() function.

make_webshot() expects the script to put the file in
$webrooturl/images/webshots, and be named $posted.png. (where $posted is
a piece of metadata, the unixtime of the post)